updated file stru.
