+++ client/src/App.css
@@ -2,6 +2,10 @@
 {
   background-color: bisque;
 }
+#ssec
+{
+    margin-top:-30px;
+}
 .mytext {
   font-weight: 600;
   font-family: 'sans-serif';
