+++ client/src/Components/Dashboard/dash.js
@@ -1,0 +1,10 @@
+import  React from "react"
+const dash = () =>
+{
+    return (
+        <div> 
+        <h>HI THERE djfh djfsd  lscdm knjfv fknj fn dfknj dfknd  fkn dfnd dfjd</h>
+        </div>
+    )
+}
+export default dash
