+++ client/src/Components/MidPage/styles/JobCategories.css
