+++ package.json
@@ -11,5 +11,8 @@
     "react-router-dom": "^5.2.0",
     "styled": "^1.0.0",
     "typescript": "^4.2.4"
+  },
+  "devDependencies": {
+    "chokidar": "^3.5.1"
   }
 }
